fix: match subcommands on aliases as well as names
previously we had run the default 'mob' subcommand
when the given arg didn't match any subcommand
Names but if the given arg matched a subcommand
Alias we didn't notice.

this commit ensures that if the first arg matches a
subcommand by name or alias that we will run that
subcommand instead of passing the arg to the mob
command as an initial.