build: add PR workflow
I am not thrilled with the way this duplicates and then
doesn't make use of the "cit" make target, since the
intention of that "cit" make target was to provide an
easy clean way to run the same build locally and on
the server.

this may be a point to investigate and refactor in the
future.

resolves #4