fix: managing coauthors does not need to be in a working tree
as the coauthors are stored globally, not locally, there is
no requirement that you be in a working tree (i.e. local git
repo/folder) to add or remove or update a coauthor from the
~/.git-coauthors file

the suggest-coauthors feature however does require that it
is in a working tree as it scans the commit history of the
current repo to suggest coauthors based on who has contributed
commits to the repo.