fix: uninstall can break hook scripts test is non-deterministic
this test would fail if go-git-mob was still in the path,
since uninstalling it would uninstall it from the locally
added bin path but not remove or hide it from the globally
installed path.

this hack adds a new matcher:

  I look for executables only in "{path}" within the current directory

which clears the PATH variable of everything except the
folder where the git CLI is installed; then prepends {path}

strictly speaking, the matcher does not quite do what the
text says because it looks for executables only in two
paths as opposed to one, and that path where git is
installed could well contain other things, but generally
speaking git will be installed into a system folder and
anyone who may contribute to go-git-mob (and run these
specs) is likely to be installing go-git-mob into a local
user directory

TL;DR this commit does not eliminate the chance of a
false positive but should greatly reduce the noise for
contributors who may forget or do not want to uninstal
go-git-mob globally before running these specs.